TestFrameworkDemo/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
colorama==0.4.6
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.1
pluggy==1.3.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.2
pytest-html==4.0.2
pytest-metadata==3.0.0
pytest-order==1.1.0
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.5
wsproto==1.2.0
